admin: moved data migration related parsing functions to separate file
Moved the data migration related utility functions to separate file to
make them reusable for topic load/unload handlers.

(cherry picked from commit b346d8d480b5f8b095d2f29ec770bb1da90de8e0)
